feat 重新生成项目基础文件

+++ src/main/java/com/dzy/river/chart/luo/writ/domain/dto/SubCategoryDTO.java
@@ -28,7 +28,7 @@ public class SubCategoryDTO implements Serializable {
     private Long id;
 
     @Schema(description = "所属主分类ID")
-    private Integer mainCategoryId;
+    private Long mainCategoryId;
 
     @Schema(description = "小分类名称")
     @NotBlank(message = "姓名不能为空")
