+++ src/main/java/com/dzy/river/chart/luo/writ/domain/resp/ContentTagResp.java
@@ -26,10 +26,10 @@ public class ContentTagResp implements Serializable {
     private Long id;
 
     @Schema(description = "内容ID")
-    private Integer contentId;
+    private Long contentId;
 
     @Schema(description = "标签ID")
-    private Integer tagId;
+    private Long tagId;
 
     @Schema(description = "删除标志：0-未删除，1-已删除")
     private Byte isDeleted;
